+++ src/features/items/data/consumableItems.ts
@@ -4,10 +4,15 @@ export const consumableItems: ConsumableItem[] = [
   {
     category: 'consumable',
     name: 'All-Seeing Eye',
-    imagePath: '/consumable/all-seeing_eye.png',
-    saveFileSlug: '',                                                           // TODO add saveFileSlug
-    id: '',                                                                     // TODO add random ID
-    tags: ['Health', 'Stamina', 'Reduce Skill Cooldown', 'Mod Power', 'All Damage'],
+    imagePath: '/consumable/all_seeing_eye.png',
+    id: 'rw5Q9E',
+    tags: [
+      'Health',
+      'Stamina',
+      'Reduce Skill Cooldown',
+      'Mod Power',
+      'All Damage',
+    ],
     description: `Grants a random buff or debuff. Lasts 3m.`,
     wikiLinks: [`https://remnant.wiki/All-Seeing_Eye`],
   },